requirements.txt
pytest
python-telegram-bot==21.9
google-genai>=1.0.0
python-dotenv==1.0.1
supabase==2.10.0
